WIP: criação da base das excpetions personalizadas, atualização nos services e criação de endereço

+++ src/main/java/com/tecnico/attornatus/pessoas/repository/PessoaDAO.java
@@ -3,6 +3,6 @@ package com.tecnico.attornatus.pessoas.repository;
 import com.tecnico.attornatus.pessoas.domain.Pessoa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PessoaDAO extends JpaRepository<Pessoa, Integer> {
+public interface PessoaDAO extends JpaRepository<Pessoa, Long> {
 
 }
